added featured content
will pull data from website.  I want it to be a collapsible title
featured then the title of the conent and once the collapsible is
clicked the details are shown but i can't get the title and details to
show up.  Also right now the title is showing up twice for some reason.